requirements.txt
langchain
fastapi
faiss-cpu
openai
PyPDF2
python-dotenv
python-multipart
uvicorn[standard]
gunicorn
networkx
youtube-transcript-api
tiktoken
passlib[bcrypt]
python-jose[cryptography]